@@ -11,19 +11,19 @@
   "private": true,
   "dependencies": {
     "@angular-material-components/color-picker": "^6.0.0",
-    "@angular/animations": "~12.2.6",
-    "@angular/cdk": "^12.2.6",
-    "@angular/common": "~12.2.6",
-    "@angular/compiler": "~12.2.6",
-    "@angular/core": "~12.2.6",
+    "@angular/animations": "~14.0.1",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "~14.0.1",
+    "@angular/compiler": "~14.0.1",
+    "@angular/core": "~14.0.1",
     "@angular/flex-layout": "^12.0.0-beta.35",
-    "@angular/forms": "~12.2.6",
-    "@angular/localize": "~12.2.6",
+    "@angular/forms": "~14.0.1",
+    "@angular/localize": "~14.0.1",
     "@angular/material": "12.2.6",
     "@angular/material-moment-adapter": "12.2.6",
-    "@angular/platform-browser": "~12.2.6",
-    "@angular/platform-browser-dynamic": "~12.2.6",
-    "@angular/router": "~12.2.6",
+    "@angular/platform-browser": "~14.0.1",
+    "@angular/platform-browser-dynamic": "~14.0.1",
+    "@angular/router": "~14.0.1",
     "@ngrx/data": "^12.4.0",
     "@ngrx/effects": "^12.4.0",
     "@ngrx/entity": "^12.4.0",
@@ -77,8 +77,8 @@
     "@angular-devkit/build-angular": "^12.2.6",
     "@angular-devkit/core": "12.2.6",
     "@angular/cli": "^12.2.6",
-    "@angular/compiler-cli": "~12.2.6",
-    "@angular/language-service": "~12.2.6",
+    "@angular/compiler-cli": "~14.0.1",
+    "@angular/language-service": "~14.0.1",
     "@types/datatables.net": "^1.10.19",
     "@types/jquery": "^3.5.1",
     "@types/node": "^12.11.1",
@@ -86,7 +86,7 @@
     "protractor": "~7.0.0",
     "ts-mxgraph": "^1.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.7.3"
   },
   "resolutions": {
     "glob-parent": "^5.1.2",
